drivers: organize clint header inclusion logic
